Added YT shorts and refactored id extraction
Joined Id extraction and video embedding functions to reduce repetition and to avoid more complicated return types from video ID extraction. The URL contains the information for whether the video in question is a YouTube short or not: youtube.com/shorts/{videoId}. Changed to youtube-nocookie.com, which seems to be a popular url for embedding youtube videos. I haven't tested this change locally since YT throws a 153 configuration error when running in such a bootleg environment.